Hide SSR import rewrite quirks behind a compatibility Adapter
SSR keeps its legacy regex scope, rewrite order, React runtime handling, and query byte ordering while its implementation moves into the import rewrite Module.

Constraint: SSR output URLs are cache keys and must remain byte-compatible

Rejected: Convert SSR to the transform strategy runner immediately | parser-driven rewrites risk whitespace and string-literal drift

Confidence: medium

Scope-risk: moderate

Directive: Keep the SSR adapter private unless a public API change is explicitly planned

Tested: RED deno test --no-check --allow-all src/transforms/import-rewriter/core.test.ts failed with missing ssr-adapter.ts

Tested: deno test --no-check --allow-all src/transforms/import-rewriter/core.test.ts src/modules/server/ssr-import-rewriter.test.ts src/modules/server/module-server.test.ts src/modules/server/module-batch-handler.test.ts

Tested: deno check src/transforms/import-rewriter/ssr-adapter.ts src/modules/server/ssr-import-rewriter.ts src/modules/server/module-server.ts src/modules/server/module-batch-handler.ts src/transforms/import-rewriter/core.test.ts

Tested: deno fmt --check src/transforms/import-rewriter/ssr-adapter.ts src/modules/server/ssr-import-rewriter.ts src/transforms/import-rewriter/core.test.ts

Tested: deno lint src/transforms/import-rewriter/ssr-adapter.ts src/modules/server/ssr-import-rewriter.ts src/transforms/import-rewriter/core.test.ts

Not-tested: Full repository test suite
